Switch to `org.creekservice`
https://github.com/creek-service/creek-test/issues/24
