+++ .github/workflows/gradle.yml
@@ -8,6 +8,9 @@ on:
     branches: [ main ]
   pull_request:
     branches: [ main ]
+  repository_dispatch:
+    types:
+      - publish_new_version
 
 jobs:
   build:
@@ -28,7 +31,7 @@ jobs:
           COVERALLS_REPO_TOKEN: ${{ secrets.COVERALLS_REPO_TOKEN }}
         run: ./gradlew check coveralls
       - name: Publish
-        if: github.event_name == 'push'
+        if: github.event_name == 'push' || github.event.action == 'publish_new_version'
         run:  |
           git fetch --tags --unshallow
           ./gradlew publish
