+++ .github/workflows/version.yml
@@ -1,0 +1,32 @@
+name: Version
+
+on:
+  workflow_dispatch:
+    inputs:
+      part:
+        description: "Part to increment: Major, Minor or Patch"
+        required: true
+        default: Patch
+
+jobs:
+  version:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up JDK 11
+        uses: actions/setup-java@v3
+        with:
+          java-version: '11'
+          distribution: 'adopt'
+          cache: gradle
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Increment Version Part
+        run: |
+          git fetch --tags --unshallow
+          ./gradlew markNextVersion -Prelease.incrementer=increment${{ github.event.inputs.part }}
+      - name: Trigger Build
+        run: |
+          curl -i -d "{\"event_type\": \"publish_new_version\"}" -H "Content-Type: application/json" -H "Accept: application/vnd.github.everest-preview+json" "https://api.github.com/repos/${{ github.repository }}/dispatches"
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
