+++ build.gradle.kts
@@ -35,7 +35,7 @@ allprojects {
     apply(plugin = "com.diffplug.spotless")
     apply(plugin = "com.github.spotbugs")
 
-    group = "org.creek"
+    group = "org.creekservice"
 
     java {
         withSourcesJar()
