+++ logging-fixtures/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module creek.observability.logging.fixtures {
     requires transitive creek.observability.logging;
 
-    exports org.creek.api.test.observability.logging.structured;
+    exports org.creekservice.api.test.observability.logging.structured;
 }
